requirements.txt
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
loguru==0.7.2 